+++ MoltenVK/MoltenVK/Utility/MVKEnvironment.h
@@ -105,9 +105,9 @@ void mvkSetConfig(const MVKConfiguration& mvkConfig);
 #   define MVK_CONFIG_SYNCHRONOUS_QUEUE_SUBMITS    mvkOSVersionIsAtLeast(MVK_CONFIG_MTLEVENT_MIN_OS)
 #endif
 
-/** Fill a Metal command buffers when each Vulkan command buffer is filled. */
+/** Fill a Metal command buffer when each Vulkan command buffer is filled. */
 #ifndef MVK_CONFIG_PREFILL_METAL_COMMAND_BUFFERS
-#   define MVK_CONFIG_PREFILL_METAL_COMMAND_BUFFERS    0
+#   define MVK_CONFIG_PREFILL_METAL_COMMAND_BUFFERS    MVK_CONFIG_PREFILL_METAL_COMMAND_BUFFERS_STYLE_NO_PREFILL
 #endif
 
 /**